Fix StyleCop warnings on Microsoft.Bot.Builder.Ai.QnA project
