+++ libraries/Microsoft.Bot.Builder.Ai.QnA/QnAMakerOptions.cs
@@ -1,0 +1,38 @@
+﻿// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+namespace Microsoft.Bot.Builder.Ai.QnA
+{
+    /// <summary>
+    /// Defines options for the QnA Maker knowledge base.
+    /// </summary>
+    public class QnAMakerOptions
+    {
+        public QnAMakerOptions()
+        {
+            ScoreThreshold = 0.3f;
+        }
+
+        /// <summary>
+        /// Gets or sets the minimum score threshold, used to filter returned results.
+        /// </summary>
+        /// <remarks>Scores are normalized to the range of 0.0 to 1.0
+        /// before filtering.</remarks>
+        /// <value>
+        /// The minimum score threshold, used to filter returned results.
+        /// </value>
+        public float ScoreThreshold { get; set; }
+
+        /// <summary>
+        /// Gets or sets the number of ranked results you want in the output.
+        /// </summary>
+        /// <value>
+        /// The number of ranked results you want in the output.
+        /// </value>
+        public int Top { get; set; }
+
+        public Metadata[] StrictFilters { get; set; }
+
+        public Metadata[] MetadataBoost { get; set; }
+    }
+}
